chore(base-driver): Restructure route groups per review feedback
Collapse the W3C and Appium route groups back into single files (W3C
spec doesn't define subgroups for the standard routes, so the split
was arbitrary), split device routes out of Appium's, and merge all
extension-spec routes (previously split between w3c/ and a "non-W3C"
bucket) into one flat extensions/ directory of 13 files, since spec
host URL doesn't determine whether a route is part of the supported
extension set.
