+++ packages/base-driver/lib/protocol/routes/appium-device.ts
