+++ packages/base-driver/lib/protocol/routes/extensions/device-posture.ts
@@ -1,10 +1,10 @@
 import type {Driver, MethodMap} from '@appium/types';
 
 /**
- * W3C Device Posture.
+ * Device Posture.
  * @see https://www.w3.org/TR/device-posture/
  */
-export const W3C_DEVICE_POSTURE_ROUTES = {
+export const DEVICE_POSTURE_ROUTES = {
   '/session/:sessionId/deviceposture': {
     POST: {command: 'setDevicePosture', payloadParams: {required: ['posture']}},
     DELETE: {command: 'clearDevicePosture'},
