+++ packages/base-driver/lib/protocol/routes/extensions/permissions.ts
@@ -1,10 +1,10 @@
 import type {Driver, MethodMap} from '@appium/types';
 
 /**
- * W3C Permissions.
+ * Permissions.
  * @see https://www.w3.org/TR/permissions/
  */
-export const W3C_PERMISSIONS_ROUTES = {
+export const PERMISSIONS_ROUTES = {
   '/session/:sessionId/permissions': {
     POST: {command: 'setPermissions', payloadParams: {required: ['descriptor', 'state']}},
   },
