+++ packages/base-driver/lib/protocol/routes/extensions/privacy.ts
@@ -1,10 +1,10 @@
 import type {Driver, MethodMap} from '@appium/types';
 
 /**
- * W3C Global Privacy Control (GPC).
+ * Global Privacy Control (GPC).
  * @see https://www.w3.org/TR/gpc/
  */
-export const W3C_PRIVACY_ROUTES = {
+export const PRIVACY_ROUTES = {
   '/session/:sessionId/privacy': {
     GET: {command: 'getGlobalPrivacyControl'},
     POST: {command: 'setGlobalPrivacyControl', payloadParams: {required: ['gpc']}},
