+++ packages/base-driver/lib/protocol/routes/extensions/reporting.ts
@@ -1,10 +1,10 @@
 import type {Driver, MethodMap} from '@appium/types';
 
 /**
- * W3C Reporting API: test report generation.
+ * Reporting API: test report generation.
  * @see https://www.w3.org/TR/reporting-1/
  */
-export const W3C_REPORTING_ROUTES = {
+export const REPORTING_ROUTES = {
   '/session/:sessionId/reporting/generate_test_report': {
     POST: {
       command: 'generateTestReport',
