--- packages/base-driver/lib/protocol/routes/w3c/window.ts
@@ -1,46 +1,0 @@
-import type {Driver, MethodMap} from '@appium/types';
-
-/**
- * W3C WebDriver: Contexts (windows and frames).
- * @see https://www.w3.org/TR/webdriver2/#contexts
- */
-export const W3C_WINDOW_ROUTES = {
-  '/session/:sessionId/window': {
-    GET: {command: 'getWindowHandle'},
-    POST: {
-      command: 'setWindow',
-      payloadParams: {
-        required: ['handle'],
-      },
-    },
-    DELETE: {command: 'closeWindow'},
-  },
-  '/session/:sessionId/window/handles': {
-    GET: {command: 'getWindowHandles'},
-  },
-  '/session/:sessionId/window/new': {
-    POST: {command: 'createNewWindow', payloadParams: {optional: ['type']}},
-  },
-  '/session/:sessionId/frame': {
-    POST: {command: 'setFrame', payloadParams: {required: ['id']}},
-  },
-  '/session/:sessionId/frame/parent': {
-    POST: {command: 'switchToParentFrame'},
-  },
-  '/session/:sessionId/window/rect': {
-    GET: {command: 'getWindowRect'},
-    POST: {
-      command: 'setWindowRect',
-      payloadParams: {optional: ['x', 'y', 'width', 'height']},
-    },
-  },
-  '/session/:sessionId/window/maximize': {
-    POST: {command: 'maximizeWindow'},
-  },
-  '/session/:sessionId/window/minimize': {
-    POST: {command: 'minimizeWindow'},
-  },
-  '/session/:sessionId/window/fullscreen': {
-    POST: {command: 'fullScreenWindow'},
-  },
-} as const satisfies MethodMap<Driver>;
